Bug 882917 - [email] "Body" search of HTML messages regressed by worker thread migration. r=mcav
land https://github.com/mozilla-b2g/gaia-email-libs-and-more/pull/297
